[CAP-722] No limits for external ach transfers should be applied
Approved-by: Buduianu George

+++ src/main/java/com/clearspend/capital/client/stripe/types/ReceivedCredit.java
@@ -16,7 +16,7 @@ public class ReceivedCredit {
   public String object;
 
   @SerializedName("amount")
-  public Integer amount;
+  public Long amount;
 
   @SerializedName("created")
   public Integer created;
